refactoring view models
