+++ Warkey.Shared/ViewModels/AutoChatViewModel.cs
@@ -1,26 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Input;
-using Warkey.Core.Items;
+using Warkey.Core.ViewModels;
 
 namespace Warkey.Core.ViewModels
 {
-    public class AutoChatViewModel : BaseViewModel
+    public class AutoChatItem : BaseViewModel
     {
-        public ObservableCollection<AutoChatItem> ListOfAutoChats { get; set; }
-
-        public Visibility ExtraCommandVisibility { get; set; }
-
-        public AutoChatViewModel()
-        {
-            ListOfAutoChats = new ObservableCollection<AutoChatItem>();
-            ExtraCommandVisibility = Visibility.Collapsed;
-        }
-
+        public Key Key { get; set; }
+        public string Message { get; set; }
     }
 }
