--- Warkey.Shared/ViewModels/EditAutoChatViewModel.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Input;
-
-namespace Warkey.Core.ViewModels
-{
-    public class EditAutoChatViewModel: BaseViewModel
-    {
-        public Key Key { get; set; }
-        public string Message { get; set; }
-    }
-}
