+++ Warkey.Shared/Warkey.Shared.csproj
@@ -42,15 +42,14 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Models\AutoChatItem.cs" />
+    <Compile Include="ViewModels\AutoChatViewModel.cs" />
     <Compile Include="Models\HotkeyItem.cs" />
     <Compile Include="Models\SettingsModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ScreenResolution.cs" />
     <Compile Include="TkokSave.cs" />
-    <Compile Include="ViewModels\AutoChatViewModel.cs" />
+    <Compile Include="ViewModels\AutoChatsViewModel.cs" />
     <Compile Include="ViewModels\BaseViewModel.cs" />
-    <Compile Include="ViewModels\EditAutoChatViewModel.cs" />
     <Compile Include="ViewModels\LoadGameViewModel.cs" />
     <Compile Include="ViewModels\MainViewModel.cs" />
     <Compile Include="ViewModels\SettingsViewModel.cs" />
